[Fixes #12076] Move the new users API under API module (#12081)
* [FIXES #12076] Move the new users API under an api module

* [FIXES #12076] Move the new users API under an api module

* [Fixes #12076] Move the new users API under API module

* [FIXES #12076] Move the new users API under an api module

* [FIXES #12076] Move the new users API under an api module
