+++ geonode/people/api/urls.py
@@ -1,0 +1,7 @@
+from geonode.people.api import views
+from geonode.api.urls import router
+
+
+router.register("users", views.UserViewSet, "users")
+
+urlpatterns = []
